Expose hidden tests for Maven
