+++ src/test/java/com/github/loadtest4j/loadtest4j/PropertiesResourceTest.java
@@ -1,0 +1,32 @@
+package com.github.loadtest4j.loadtest4j;
+
+import com.github.loadtest4j.loadtest4j.junit.IntegrationTest;
+import org.junit.Test;
+import org.junit.experimental.categories.Category;
+
+import java.util.Properties;
+
+import static org.junit.Assert.assertEquals;
+
+@Category(IntegrationTest.class)
+public class PropertiesResourceTest {
+    @Test
+    public void getProperties() {
+        final LoadTesterFactory.PropertiesResource sut = new LoadTesterFactory.PropertiesResource("/props/example.properties");
+
+        final Properties properties = sut.getProperties();
+
+        final Properties expectedProperties = new Properties();
+        expectedProperties.put("foo", "bar");
+        assertEquals(expectedProperties, properties);
+    }
+
+    @Test
+    public void getEmptyProperties() {
+        final LoadTesterFactory.PropertiesResource sut = new LoadTesterFactory.PropertiesResource("/props/fake.properties");
+
+        final Properties properties = sut.getProperties();
+
+        assertEquals(new Properties(), properties);
+    }
+}
